Add new resource for function state
