+++ streampipes-commons/src/main/java/org/apache/streampipes/commons/constants/Envs.java
@@ -48,6 +48,7 @@ public enum Envs {
   SP_OAUTH_ENABLED("SP_OAUTH_ENABLED", "false"),
   SP_OAUTH_REDIRECT_URI("SP_OAUTH_REDIRECT_URI"),
   SP_RESET_ENDPOINT_ENABLED("SP_RESET_ENDPOINT_ENABLED", "false"),
+  SP_FUNCTION_STATE_ENDPOINTS_ENABLED("SP_FUNCTION_STATE_ENDPOINTS_ENABLED", "false"),
   SP_DEBUG("SP_DEBUG", "false"),
   SP_MAX_WAIT_TIME_AT_SHUTDOWN("SP_MAX_WAIT_TIME_AT_SHUTDOWN"),
