+++ streampipes-rest/src/main/java/org/apache/streampipes/rest/impl/FunctionStateResource.java
@@ -1,0 +1,93 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package org.apache.streampipes.rest.impl;
+
+import org.apache.streampipes.model.function.FunctionState;
+import org.apache.streampipes.model.message.Notifications;
+import org.apache.streampipes.model.message.SuccessMessage;
+import org.apache.streampipes.rest.core.base.impl.AbstractAuthGuardedRestResource;
+import org.apache.streampipes.rest.shared.exception.SpMessageException;
+import org.apache.streampipes.storage.api.function.IFunctionStateStorage;
+
+import org.springframework.context.annotation.Conditional;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.Map;
+
+@RestController
+@RequestMapping("/api/v2/functions")
+@Conditional(FunctionStateEndpointsEnabledCondition.class)
+public class FunctionStateResource extends AbstractAuthGuardedRestResource {
+
+  private final IFunctionStateStorage functionStateStorage;
+
+  public FunctionStateResource(IFunctionStateStorage functionStateStorage) {
+    this.functionStateStorage = functionStateStorage;
+  }
+
+  @GetMapping(path = "{functionId}/state", produces = MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<Map<String, Object>> getFunctionState(@PathVariable("functionId") String functionId) {
+    var functionState = functionStateStorage.getElementById(functionId);
+    if (functionState != null) {
+      return ok(functionState.getState());
+    } else {
+      throw new SpMessageException(HttpStatus.NOT_FOUND, Notifications.error("Function state not found"));
+    }
+  }
+
+  @PutMapping(
+      path = "{functionId}/state",
+      produces = MediaType.APPLICATION_JSON_VALUE,
+      consumes = MediaType.APPLICATION_JSON_VALUE
+  )
+  public ResponseEntity<SuccessMessage> persistFunctionState(@PathVariable("functionId") String functionId,
+                                                             @RequestBody Map<String, Object> state) {
+    var existingFunctionState = functionStateStorage.getElementById(functionId);
+
+    if (existingFunctionState != null) {
+      existingFunctionState.setState(state);
+      functionStateStorage.updateElement(existingFunctionState);
+    } else {
+      functionStateStorage.persist(new FunctionState(functionId, state));
+    }
+
+    return ok(Notifications.success("Function state successfully persisted"));
+  }
+
+  @DeleteMapping(path = "{functionId}/state", produces = MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<SuccessMessage> deleteFunctionState(@PathVariable("functionId") String functionId) {
+    var existingFunctionState = functionStateStorage.getElementById(functionId);
+
+    if (existingFunctionState == null) {
+      throw new SpMessageException(HttpStatus.NOT_FOUND, Notifications.error("Function state not found"));
+    }
+
+    functionStateStorage.deleteElementById(functionId);
+    return ok(Notifications.success("Function state successfully deleted"));
+  }
+}
