+++ streampipes-rest/src/main/java/org/apache/streampipes/rest/impl/FunctionsResource.java
@@ -21,17 +21,13 @@ package org.apache.streampipes.rest.impl;
 import org.apache.streampipes.loadbalance.pipeline.ExtensionsLogProvider;
 import org.apache.streampipes.manager.function.FunctionRegistrationService;
 import org.apache.streampipes.model.function.FunctionDefinition;
-import org.apache.streampipes.model.function.FunctionState;
 import org.apache.streampipes.model.message.Notifications;
 import org.apache.streampipes.model.message.SuccessMessage;
 import org.apache.streampipes.model.monitoring.SpLogEntry;
 import org.apache.streampipes.model.monitoring.SpMetricsEntry;
 import org.apache.streampipes.rest.core.base.impl.AbstractAuthGuardedRestResource;
 import org.apache.streampipes.rest.security.AuthConstants;
-import org.apache.streampipes.rest.shared.exception.SpMessageException;
-import org.apache.streampipes.storage.api.function.IFunctionStateStorage;
 
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -39,24 +35,16 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.Collection;
 import java.util.List;
-import java.util.Map;
 
 @RestController
 @RequestMapping("/api/v2/functions")
 public class FunctionsResource extends AbstractAuthGuardedRestResource {
-
-  private final IFunctionStateStorage functionStateStorage;
-
-  public FunctionsResource(IFunctionStateStorage functionStateStorage) {
-    this.functionStateStorage = functionStateStorage;
-  }
 
   @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
   @PreAuthorize(AuthConstants.IS_ADMIN_ROLE)
@@ -97,46 +85,5 @@ public class FunctionsResource extends AbstractAuthGuardedRestResource {
   @PreAuthorize(AuthConstants.IS_ADMIN_ROLE)
   public ResponseEntity<List<SpLogEntry>> getFunctionLogs(@PathVariable("functionId") String functionId) {
     return ok(ExtensionsLogProvider.INSTANCE.getLogInfosForResource(functionId));
-  }
-
-  @GetMapping(path = "{functionId}/state", produces = MediaType.APPLICATION_JSON_VALUE)
-  public ResponseEntity<Map<String, Object>> getFunctionState(@PathVariable("functionId") String functionId) {
-    var functionState = functionStateStorage.getElementById(functionId);
-    if (functionState != null) {
-      return ok(functionState.getState());
-    } else {
-      throw new SpMessageException(HttpStatus.NOT_FOUND, Notifications.error("Function state not found"));
-    }
-  }
-
-  @PutMapping(
-      path = "{functionId}/state",
-      produces = MediaType.APPLICATION_JSON_VALUE,
-      consumes = MediaType.APPLICATION_JSON_VALUE
-  )
-  public ResponseEntity<SuccessMessage> persistFunctionState(@PathVariable("functionId") String functionId,
-                                                             @RequestBody Map<String, Object> state) {
-    var existingFunctionState = functionStateStorage.getElementById(functionId);
-
-    if (existingFunctionState != null) {
-      existingFunctionState.setState(state);
-      functionStateStorage.updateElement(existingFunctionState);
-    } else {
-      functionStateStorage.persist(new FunctionState(functionId, state));
-    }
-
-    return ok(Notifications.success("Function state successfully persisted"));
-  }
-
-  @DeleteMapping(path = "{functionId}/state", produces = MediaType.APPLICATION_JSON_VALUE)
-  public ResponseEntity<SuccessMessage> deleteFunctionState(@PathVariable("functionId") String functionId) {
-    var existingFunctionState = functionStateStorage.getElementById(functionId);
-
-    if (existingFunctionState == null) {
-      throw new SpMessageException(HttpStatus.NOT_FOUND, Notifications.error("Function state not found"));
-    }
-
-    functionStateStorage.deleteElementById(functionId);
-    return ok(Notifications.success("Function state successfully deleted"));
   }
 }
